+++ lib/dataset/include/scipp/dataset/map_view.h
@@ -188,4 +188,13 @@ template <class Masks>
 template <class Key, class Value>
 bool equals_nan(const Dict<Key, Value> &a, const Dict<Key, Value> &b);
 
+constexpr auto get_data = [](auto &&a) -> decltype(auto) { return a.data(); };
+constexpr auto get_sizes = [](auto &&a) -> decltype(auto) { return a.sizes(); };
+constexpr auto get_meta = [](auto &&a) -> decltype(auto) { return a.meta(); };
+constexpr auto get_coords = [](auto &&a) -> decltype(auto) {
+  return a.coords();
+};
+constexpr auto get_attrs = [](auto &&a) -> decltype(auto) { return a.attrs(); };
+constexpr auto get_masks = [](auto &&a) -> decltype(auto) { return a.masks(); };
+
 } // namespace scipp::dataset