+++ src/scipp/__init__.py
@@ -105,8 +105,8 @@ del _cls
 del _binding
 
 from . import data
-
 from . import serialization
+from . import spatial
 
 from .plotting import plot