+++ src/scipp/core/cpp_classes.py
@@ -1,0 +1,12 @@
+# SPDX-License-Identifier: BSD-3-Clause
+# Copyright (c) 2022 Scipp contributors (https://github.com/scipp)
+
+# flake8: noqa: F401
+
+from .._scipp.core import Coords, DataArray, Dataset, DType, GroupByDataArray, \
+    GroupByDataset, Masks, Unit, Variable
+
+from .._scipp.core import BinEdgeError, BinnedDataError, CoordError, \
+    DataArrayError, DatasetError, DimensionError, \
+    DTypeError, SizeError, SliceError, \
+    UnitError, VariableError, VariancesError